Fix clippy

+++ crates/cubecl-core/src/frontend/container/array/base.rs
@@ -326,7 +326,7 @@ impl<T: CubePrimitive> ListExpand<T> for ExpandElementTyped<Array<T>> {
     }
 
     fn line_size(&self) -> u32 {
-        self.expand.ty.line_size() as u32
+        self.expand.ty.line_size()
     }
 }
