+++ crates/cubecl-spirv/src/subgroup.rs
@@ -35,7 +35,7 @@ impl<T: SpirvTarget> SpirvCompiler<T> {
                             let ids = (0..vec)
                                 .map(|i| {
                                     let elem_i = b
-                                        .composite_extract(elem_ty, None, input, vec![i as u32])
+                                        .composite_extract(elem_ty, None, input, vec![i])
                                         .unwrap();
                                     b.group_non_uniform_all(bool_ty, None, subgroup, elem_i)
                                         .unwrap()
@@ -63,7 +63,7 @@ impl<T: SpirvTarget> SpirvCompiler<T> {
                             let ids = (0..vec)
                                 .map(|i| {
                                     let elem_i = b
-                                        .composite_extract(elem_ty, None, input, vec![i as u32])
+                                        .composite_extract(elem_ty, None, input, vec![i])
                                         .unwrap();
                                     b.group_non_uniform_any(bool_ty, None, subgroup, elem_i)
                                         .unwrap()
